Revert magma_int_t to int (not long)
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@2059 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ include/magmablas.h
@@ -12,7 +12,7 @@
 #include <cublas.h>
 #include <cuda.h>
 
-typedef long magma_int_t;
+typedef int  magma_int_t;
 typedef int  magma_err_t;
 typedef void* magma_devptr;
